linux: Move `mount.h` API to `src/new`
This has the effect of making `MOVE_MOUNT_*` constants available on all
Linux platforms, not just glibc.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/uapi/linux/mount.h?h=028ef9c96e96197026887c0f092424679298aae8
