+++ src/new/linux_uapi/linux/mod.rs
@@ -5,5 +5,6 @@
 pub(crate) mod can;
 pub(crate) mod keyctl;
 pub(crate) mod membarrier;
+pub(crate) mod mount;
 pub(crate) mod netlink;
 pub(crate) mod pidfd;
