+++ src/new/mod.rs
@@ -182,6 +182,7 @@ cfg_if! {
         pub use linux::can::raw::*;
         pub use linux::keyctl::*;
         pub use linux::membarrier::*;
+        pub use linux::mount::*;
         pub use linux::netlink::*;
         pub use linux::pidfd::*;
         #[cfg(target_env = "gnu")]
